Refactor the Layout
Refactor the layout to allow style sheet type drawing
mode independently from the widgets themselves.
